+++ admin/plugin-meta.php
@@ -14,6 +14,8 @@ class Fact_Maven_Disable_Blogging {
         add_filter( 'plugin_row_meta', array( $this, 'plugin_row_meta' ), 10, 2 );
         # Add link to plugin settings
         add_filter( 'plugin_action_links', array( $this, 'plugin_action_links' ), 10, 2 );
+        # Handle localisation
+        add_action( 'plugins_loaded', array( $this, 'i18n' ), 0 );
     }
 
     public function plugin_row_meta( $links, $file ) {
@@ -40,6 +42,11 @@ class Fact_Maven_Disable_Blogging {
         }
         # Return the settings link
         return $links;
+    }
+
+    public function i18n() {
+        # Load the translation of the plugin
+        load_plugin_textdomain( 'dsbl', false, basename( dirname( __FILE__ ) ) . '/languages/' );
     }
 }